IK-0010 2. Создание сущности ЗПК и замена всех id с Integer на Long

+++ Slam/src/main/java/com/system/slam/entity/Autor.java
@@ -14,7 +14,7 @@ public class Autor {
     @Id
     @Column(name = "IdAutor", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idAutor;
+    private Long idAutor;
 
     @Column(name = "LastName", length = 20, nullable = false)
     private String lastName;
